add motion blur
a time parameter was added to the ray to record changes in time.

this, coupled with a movement of the sphere means that we can
calculate motion by moving the sphere and seeing where does the ray
interact with the sphere.

one idea to improve the picture can be to sample the sphere in a
smaller time range (roughly equal to a single frame of a camera shutter)